add the irqwatch command
`irqwatch` is meant to help debug/troubleshooting irq pinning scenarios.
We monitor `/proc/interrupts` and we report if a given cpu set handled
IRQs in a given time interval.

With proper IRQ pinning set, the isolated cpus should never handle
IRQs at all; unfortunately, the only way to check this fact is ex-post,
so we need to monitor cpus for long-enough time interval.

Signed-off-by: Francesco Romani <fromani@redhat.com>